requirements.txt
streamlit==1.32.0
pandas==2.2.1
fpdf2==2.7.9